Rename extension-polyfill to browser-polyfill and move to browser directory
