+++ app/scripts/contentscript.js
@@ -6,7 +6,7 @@ import { obj as createThoughStream } from 'through2';
 
 import { isManifestV3 } from '../../shared/modules/mv3.utils';
 import shouldInjectProvider from '../../shared/modules/provider-injection';
-import { browser } from './desktop/extension-polyfill';
+import { browser } from './desktop/browser/browser-polyfill';
 
 // These require calls need to use require to be statically recognized by browserify
 const fs = require('fs');
