+++ app/scripts/detect-multiple-instances.js
@@ -11,7 +11,7 @@ import {
   CHROME_BUILD_IDS,
   FIREFOX_BUILD_IDS,
 } from '../../shared/constants/app';
-import { browser } from './desktop/extension-polyfill';
+import { browser } from './desktop/browser/browser-polyfill';
 import { getPlatform } from './lib/util';
 
 const MESSAGE_TEXT = 'isRunning';
