+++ app/scripts/lib/get-first-preferred-lang-code.js
@@ -1,4 +1,4 @@
-import { browser } from '../desktop/extension-polyfill';
+import { browser } from '../desktop/browser/browser-polyfill';
 import allLocales from '../../_locales/index.json';
 
 // mapping some browsers return hyphen instead underscore in locale codes (e.g. zh_TW -> zh-tw)
