+++ shared/modules/mv3.utils.js
@@ -1,5 +1,5 @@
 import cfg from '../../app/scripts/desktop/config';
-import { browser } from '../../app/scripts/desktop/extension-polyfill';
+import { browser } from '../../app/scripts/desktop/browser/browser-polyfill';
 
 export const isManifestV3 =
   browser.runtime.getManifest().manifest_version === 3 ||
