+++ ui/components/app/menu-bar/menu-bar.js
@@ -1,7 +1,7 @@
 import React, { useState, useContext } from 'react';
 import { useHistory } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { browser } from '../../../../app/scripts/desktop/extension-polyfill';
+import { browser } from '../../../../app/scripts/desktop/browser/browser-polyfill';
 import SelectedAccount from '../selected-account';
 import ConnectedStatusIndicator from '../connected-status-indicator';
 import { getEnvironmentType } from '../../../../app/scripts/lib/util';
